+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.50] - 2025-01-xx
 
 - [#23](https://github.com/estruyf/vscode-demo-time/issues/23): Add the ability to add notes to demos
+- [#24](https://github.com/estruyf/vscode-demo-time/issues/24): Added the `close` and `closeAll` actions to close editors
 
 ## [0.0.49] - 2024-12-30